Revert "Revert "Revert "Revert "Revert "Merge branch 'overworld-expansion' of https://github.com/ghoulslash/pokeemerald"""""
This reverts commit af639a3c9d16f6bc572feeaea8a5acf4bf6038a5.

+++ src/easy_chat.c
@@ -4988,7 +4988,7 @@ static void HideStartSelectButtons(void)
 
 static void TryAddInterviewObjectEvents(void)
 {
-    u16 graphicsId;
+    int graphicsId;
     u8 spriteId;
 
     switch (GetDisplayedPersonType())
